+++ src/app.ts
@@ -1,11 +1,13 @@
 process.on('unhandledRejection', console.dir);
 
-// モジュールの取り込み
-import config from 'config';
-import { CronJob } from 'cron';
-import http from 'http';
+import { AndroidApp, IosApp } from 'application';
 
-import Review from './domains/Review';
+import { CronJob } from 'cron';
+import { androidReview } from './domains/Android';
+import { changeToArray } from './utils/array';
+import config from 'config';
+import http from 'http';
+import { iosReview } from './domains/Ios';
 
 const CRON_TIME: string = config.get('cron.time');
 const TIME_ZONE: string = config.get('cron.timeZone');
@@ -44,8 +46,11 @@ if (config.get('maxConnections')) {
 }
 
 try {
-//  const CronJob = require('cron').CronJob;
-  new CronJob(CRON_TIME, function() {
+  const iosApps: IosApp[] = changeToArray(config.get('app.iOS'));
+  const androidApps: AndroidApp[] = changeToArray(config.get('app.android'));
+
+  //  const CronJob = require('cron').CronJob;
+  new CronJob(CRON_TIME, function () {
 
     // 未設定の場合は全件表示
     if (outputs === null) {
@@ -56,11 +61,20 @@ try {
     //   outputs = parseInt(outputs);
     // }
 
-    try {
-      let app = new Review(outputs, ignoreNotification, config, db, useSlack, useEmail);
-      app.main();
-    } catch (e) {
-      console.log("application error: " + e);
+    if (iosApps) {
+      try {
+        iosReview({ iosApps, outputs, ignoreNotification, db, useSlack, useEmail });
+      } catch (e) {
+        console.log("application error(iOS): " + e);
+      }
+    }
+
+    if (androidApps) {
+      try {
+        androidReview({ androidApps, outputs, ignoreNotification, db, useSlack, useEmail });
+      } catch (e) {
+        console.log("application error(Android): " + e);
+      }
     }
 
     // 通知しない設定をオフにする
